refactor(app, components): migrate useIdle to app
Because we will need to special-case the SLEEP_NEVER_MS as being a case in which we never idle, this
util should live in the app
