+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -9,7 +9,6 @@ import {
   COLORS,
   OVERFLOW_AUTO,
   POSITION_RELATIVE,
-  useIdle,
   useScrolling,
 } from '@opentrons/components'
 import { ApiHostProvider } from '@opentrons/react-api-client'
@@ -52,7 +51,7 @@ import {
   updateConfigValue,
 } from '/app/redux/config'
 import { updateBrightness } from '/app/redux/shell'
-import { SLEEP_NEVER_MS } from '/app/local-resources/config'
+import { useIdle, SLEEP_NEVER_MS } from '/app/local-resources/dom-utils'
 import { useProtocolReceiptToast, useSoftwareUpdatePoll } from './hooks'
 import { ODDTopLevelRedirects } from './ODDTopLevelRedirects'
 import { ReactQueryDevtools } from '/app/App/tools'
