+++ app/src/local-resources/dom-utils/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useScrollPosition'
+export * from './useIdle'
